+++ analysis/src/Hover.ml
@@ -101,26 +101,50 @@ let findRelevantTypesFromType ~file ~package typ =
   constructors |> List.filter_map (fromConstructorPath ~env:envToSearch)
 
 let expandTypes ~file ~package ~supportsMarkdownLinks typ =
-  findRelevantTypesFromType typ ~file ~package
-  |> List.map (fun {decl; env; loc; path} ->
-         let linkToTypeDefinitionStr =
-           if supportsMarkdownLinks then
-             Markdown.goToDefinitionText ~env ~pos:loc.Warnings.loc_start
-           else ""
-         in
-         Markdown.divider
-         ^ (if supportsMarkdownLinks then Markdown.spacing else "")
-         ^ Markdown.codeBlock
-             (decl
-             |> Shared.declToString ~printNameAsIs:true
-                  (SharedTypes.pathIdentToString path))
-         ^ linkToTypeDefinitionStr ^ "\n")
+  match findRelevantTypesFromType typ ~file ~package with
+  | {decl; path} :: _
+    when Res_parsetree_viewer.has_inline_record_definition_attribute
+           decl.type_attributes ->
+    (* We print inline record types just with their definition, not the constr pointing 
+    to them, since that doesn't make sense to show the user. *)
+    ( [
+        Markdown.codeBlock
+          (decl
+          |> Shared.declToString ~printNameAsIs:true
+               (SharedTypes.pathIdentToString path));
+      ],
+      `InlineType )
+  | all ->
+    ( all
+      |> List.map (fun {decl; env; loc; path} ->
+             let linkToTypeDefinitionStr =
+               if
+                 supportsMarkdownLinks
+                 && not
+                      (Res_parsetree_viewer
+                       .has_inline_record_definition_attribute
+                         decl.type_attributes)
+               then Markdown.goToDefinitionText ~env ~pos:loc.Warnings.loc_start
+               else ""
+             in
+             Markdown.divider
+             ^ (if supportsMarkdownLinks then Markdown.spacing else "")
+             ^ Markdown.codeBlock
+                 (decl
+                 |> Shared.declToString ~printNameAsIs:true
+                      (SharedTypes.pathIdentToString path))
+             ^ linkToTypeDefinitionStr ^ "\n"),
+      `Default )
 
 (* Produces a hover with relevant types expanded in the main type being hovered. *)
 let hoverWithExpandedTypes ~file ~package ~supportsMarkdownLinks typ =
   let typeString = Markdown.codeBlock (typ |> Shared.typeToString) in
-  typeString :: expandTypes ~file ~package ~supportsMarkdownLinks typ
-  |> String.concat "\n"
+  let expandedTypes, expansionType =
+    expandTypes ~file ~package ~supportsMarkdownLinks typ
+  in
+  match expansionType with
+  | `Default -> typeString :: expandedTypes |> String.concat "\n"
+  | `InlineType -> expandedTypes |> String.concat "\n"
 
 (* Leverages autocomplete functionality to produce a hover for a position. This
    makes it (most often) work with unsaved content. *)
@@ -171,10 +195,13 @@ let newHover ~full:{file; package} ~supportsMarkdownLinks locItem =
     let typeDef = Markdown.codeBlock (Shared.declToString name decl) in
     match decl.type_manifest with
     | None -> Some typeDef
-    | Some typ ->
-      Some
-        (typeDef :: expandTypes ~file ~package ~supportsMarkdownLinks typ
-        |> String.concat "\n"))
+    | Some typ -> (
+      let expandedTypes, expansionType =
+        expandTypes ~file ~package ~supportsMarkdownLinks typ
+      in
+      match expansionType with
+      | `Default -> Some (typeDef :: expandedTypes |> String.concat "\n")
+      | `InlineType -> Some (expandedTypes |> String.concat "\n")))
   | LModule (Definition (stamp, _tip)) | LModule (LocalReference (stamp, _tip))
     -> (
     match Stamps.findModule file.stamps stamp with