+++ compiler/ml/types.mli
@@ -262,7 +262,12 @@ type type_declaration = {
   type_attributes: Parsetree.attributes;
   type_immediate: bool; (* true iff type should not be a pointer *)
   type_unboxed: unboxed_status;
+  type_inlined_types: type_inlined_type list;
+      (** Representation of inlined types, needed for printing *)
 }
+
+and type_inlined_type =
+  | Record of {type_name: string; labels: label_declaration list}
 
 and type_kind =
   | Type_abstract