+++ compiler/syntax/src/res_parsetree_viewer.ml
@@ -72,6 +72,13 @@ let has_await_attribute attrs =
       | _ -> false)
     attrs
 
+let has_inline_record_definition_attribute attrs =
+  List.exists
+    (function
+      | {Location.txt = "res.inlineRecordDefinition"}, _ -> true
+      | _ -> false)
+    attrs
+
 let has_res_pat_variant_spread_attribute attrs =
   List.exists
     (function
@@ -198,7 +205,8 @@ let filter_parsing_attrs attrs =
             Location.txt =
               ( "res.braces" | "ns.braces" | "res.iflet" | "res.ternary"
               | "res.await" | "res.template" | "res.taggedTemplate"
-              | "res.patVariantSpread" | "res.dictPattern" );
+              | "res.patVariantSpread" | "res.dictPattern"
+              | "res.inlineRecordDefinition" );
           },
           _ ) ->
         false
@@ -352,7 +360,7 @@ let has_attributes attrs =
       | ( {
             Location.txt =
               ( "res.braces" | "ns.braces" | "res.iflet" | "res.ternary"
-              | "res.await" | "res.template" );
+              | "res.await" | "res.template" | "res.inlineRecordDefinition" );
           },
           _ ) ->
         false
@@ -547,7 +555,7 @@ let is_printable_attribute attr =
   | ( {
         Location.txt =
           ( "res.iflet" | "res.braces" | "ns.braces" | "JSX" | "res.await"
-          | "res.template" | "res.ternary" );
+          | "res.template" | "res.ternary" | "res.inlineRecordDefinition" );
       },
       _ ) ->
     false